refactor: creating half section and refactoring code for better server use
