+++ src/app/(half)/layout.tsx
@@ -1,0 +1,5 @@
+const Layout = ({ children }: { children: React.ReactNode }) => {
+  return <div className="w-[55%]">{children}</div>;
+};
+
+export default Layout;
