--- src/app/post/create/page.tsx
@@ -1,13 +1,0 @@
-import { CreatePostForm } from "./page.client";
-
-const Page = () => {
-  return (
-    <div className="flex justify-center w-full">
-      <div className="w-[55%]">
-        <CreatePostForm />
-      </div>
-    </div>
-  );
-};
-
-export default Page;
